+++ example_userdata_stream.py
@@ -67,22 +67,31 @@ binance_com_api_secret = ""
 binance_je_api_key = ""
 binance_je_api_secret = ""
 
+# configure api key and secret for binance.us
+binance_us_api_key = ""
+binance_us_api_secret = ""
+
 # create instances of BinanceWebSocketApiManager
 binance_com_websocket_api_manager = BinanceWebSocketApiManager(exchange="binance.com")
 binance_je_websocket_api_manager = BinanceWebSocketApiManager(exchange="binance.je")
+binance_us_websocket_api_manager = BinanceWebSocketApiManager(exchange="binance.us")
 
 # set api key and secret in api manager
 binance_com_websocket_api_manager.set_private_api_config(binance_com_api_key, binance_com_api_secret)
 binance_je_websocket_api_manager.set_private_api_config(binance_je_api_key, binance_je_api_secret)
+binance_us_websocket_api_manager.set_private_api_config(binance_us_api_key, binance_us_api_secret)
 
 # create the userData streams
 binance_com_user_data_stream_id = binance_com_websocket_api_manager.create_stream('arr', '!userData')
 binance_je_user_data_stream_id = binance_je_websocket_api_manager.create_stream('arr', '!userData')
+binance_us_user_data_stream_id = binance_us_websocket_api_manager.create_stream('arr', '!userData')
 
 # start a worker process to move the received stream_data from the stream_buffer to a print function
 worker_thread = threading.Thread(target=print_stream_data_from_stream_buffer, args=(binance_com_user_data_stream_id,))
 worker_thread.start()
 worker_thread = threading.Thread(target=print_stream_data_from_stream_buffer, args=(binance_je_user_data_stream_id,))
+worker_thread.start()
+worker_thread = threading.Thread(target=print_stream_data_from_stream_buffer, args=(binance_us_user_data_stream_id,))
 worker_thread.start()
 
 # monitor the streams
@@ -90,4 +99,5 @@ while True:
     binance_com_websocket_api_manager.print_stream_info(binance_com_user_data_stream_id)
     binance_com_websocket_api_manager.print_summary()
     binance_je_websocket_api_manager.print_summary()
+    binance_us_websocket_api_manager.print_summary()
     time.sleep(1)