+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to 
 [Semantic Versioning](http://semver.org/).
 
-## 1.6.6.dev (development stage/unreleased)
+## 1.7.0.dev (development stage/unreleased)
+
+## 1.7.0
+## Added 
+- binance.us websocket support
 
 ## 1.6.6
 ### Fixed